+++ assets/videos.json
@@ -76,10 +76,6 @@
         "title": "Greenpress Overview / David Meir-Levy [Pull Request]"
       },
       {
-        "src": "KuzSu4y2fJQ",
-        "title": "Objectbuffer Overview / Bnaya Perez [Pull Request]"
-      },
-      {
         "src": "JI_QTB_JT2A",
         "title": "Mini Services / Snir Shecther [Pull Request]"
       },
