--- components/talksPage/carousel/carousel.module.scss
@@ -1,67 +1,0 @@
-$itemGrow: 1.2;
-$duration: 250ms;
-
-.title {
-  font-weight: 800;
-  font-size: 36px;
-  color: #39475d;
-  line-height: 47px;
-}
-
-.wrapper {
-  display: grid;
-  grid-template-columns: repeat(4, 100%);
-  overflow: hidden;
-  height: 200px;
-  scroll-behavior: smooth;
-
-  .section {
-    width: 100%;
-    position: relative;
-    display: grid;
-    grid-template-columns: repeat(5, auto);
-    margin: 20px 0;
-
-    .item {
-      padding: 0 2px;
-      transition: $duration all;
-
-      &:hover {
-        margin: 0 40px;
-        transform: scale(1.2);
-      }
-    }
-
-    .arrow__btn {
-      position: absolute;
-      color: #fff;
-      text-decoration: none;
-      font-size: 6em;
-      background: rgb(0, 0, 0);
-      width: 80px;
-      padding: 20px;
-      text-align: center;
-      z-index: 1;
-
-      &:nth-of-type(1) {
-        top: 0;
-        bottom: 0;
-        left: 0;
-        background: linear-gradient(-90deg, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 1) 100%);
-      }
-      &:nth-of-type(2) {
-        top: 0;
-        bottom: 0;
-        right: 0;
-        background: linear-gradient(90deg, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 1) 100%);
-      }
-    }
-  }
-}
-
-// Remove the arrow for Mobile
-@media only screen and (max-width: 600px) {
-  .arrow__btn {
-    display: none;
-  }
-}
